Changed for my ec2 instance - requires some changes in settings.py and mysql database swimcoach

+++ meets/models.py
@@ -50,6 +50,6 @@ class Lane(models.Model):
 class PhoneNumber(models.Model):
 	person = models.ForeignKey(Person)
 	owner_name = models.CharField(max_length=200)
-	number = models.IntegerField(max_length=11)
+	number = models.BigIntegerField(max_length=11)
 	def __unicode__(self):
 		return str(self.number)
